+++ SkillHelper/src/main/kotlin/com/skillhelper/repository/models/SkillDbo.kt
@@ -6,6 +6,6 @@ data class SkillDbo(
     val description: String,
     val stressLevel: Int,
     val author: String?,
-    val visibility: Int,
+    val visibility: Long,
     val imageSrc: String?
 )
